+++ src/test/java/org/team1/keyduck/bidding/service/BiddingServiceTest.java
@@ -1,26 +1,17 @@
 package org.team1.keyduck.bidding.service;
 
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-import org.junit.platform.commons.logging.Logger;
-import org.junit.platform.commons.logging.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.annotation.Commit;
-import org.springframework.test.context.ActiveProfiles;
-import org.team1.keyduck.auction.entity.Auction;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.team1.keyduck.auction.repository.AuctionRepository;
 import org.team1.keyduck.auth.entity.AuthMember;
 import org.team1.keyduck.bidding.dto.response.BiddingResponseDto;
@@ -29,402 +20,164 @@ import org.team1.keyduck.bidding.repository.BiddingRepository;
 import org.team1.keyduck.common.exception.DataInvalidException;
 import org.team1.keyduck.common.exception.OperationNotAllowedException;
 import org.team1.keyduck.keyboard.repository.KeyboardRepository;
-import org.team1.keyduck.member.entity.Member;
-import org.team1.keyduck.member.entity.MemberRole;
 import org.team1.keyduck.member.repository.MemberRepository;
 import org.team1.keyduck.payment.repository.PaymentDepositRepository;
 import org.team1.keyduck.payment.service.PaymentDepositService;
 import org.team1.keyduck.testdata.TestData;
 
-@SpringBootTest
-@ActiveProfiles("test")
-public class BiddingServiceTest {
 @ExtendWith(MockitoExtension.class)
-@MockitoSettings(strictness = Strictness.LENIENT)
-class BiddingServiceTest {
-
-    @Autowired
-    private BiddingService biddingService;
-    @Autowired
-    private BiddingRepository biddingRepository;
-    @Autowired
-    @Mock
-    private MemberRepository memberRepository;
-    @Mock
-    private AuctionRepository auctionRepository;
-    @Autowired
-    private MemberRepository memberRepository;
-    @Autowired
-    private KeyboardRepository keyboardRepository;
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-    @Mock
-    private BiddingRepository biddingRepository;
-    @Mock
-    private PaymentDepositService paymentDepositService;
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+public class BiddingServiceTest {
 
     @InjectMocks
     private BiddingService biddingService;
-
-    @Test
-    void createBidding_success_첫번쨰() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID1;
-        Long memberId = TestData.TEST_ID2;
-        Long price = 25000L;
-
-    private Member member;
-    private Auction auction;
-        AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
-
-    private static final Logger log = LoggerFactory.getLogger(BiddingServiceTest.class);
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-    public void setUpBidding() {
-        Auction auction = TestData.TEST_AUCTION1; //  실제 객체 사용
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        // 판매자 멤버 생성
-        member = memberRepository.save(
-                Member.builder()
-                        .name("판매자")
-                        .email("email@example.com")
-                        .password("1234")
-                        .memberRole(MemberRole.SELLER)
-                        .address(new Address("경기도", "안양시", "동안구", "12345", "1층"))
-                        .build()
-        );
-        when(biddingRepository.findByMember_IdAndAuction_Id(memberId, auctionId)).thenReturn(
-                Optional.empty());
-
-        // 키보드 생성
-        Keyboard keyboard = keyboardRepository.save(
-                Keyboard.builder()
-                        .member(member)
-                        .name("키보드")
-                        .description("키보드입니다.")
-                        .build()
-        );
-
-        // 경매 생성
-        auction = auctionRepository.save(
-                Auction.builder()
-                        .title("키보드 경매")
-                        .keyboard(keyboard)
-                        .startPrice(1000L)
-                        .currentPrice(1000L)
-                        .biddingUnit(100L)
-                        .auctionStartDate(LocalDateTime.now())
-                        .auctionEndDate(LocalDateTime.now().plusDays(1))
-                        .auctionStatus(AuctionStatus.IN_PROGRESS)
-                        .build()
-        );
-
-        Faker faker = new Faker();
-        final int MAX_CUSTOMERS = 100;
-        final int BATCH_SIZE = 100;
-        final String PASSWORD = "5678";
-
-        String memberSql = "INSERT INTO member (name, email, password, member_role, city, state, street, detail_address1, detail_address2) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)";
-        List<Object[]> members = new ArrayList<>();
-
-        for (int i = 0; i < MAX_CUSTOMERS; i++) {
-            members.add(new Object[]{
-                    "구매자" + i,
-                    faker.color().name() + i + "@" + faker.animal() + ".com",
-                    PASSWORD,
-                    MemberRole.CUSTOMER.name(),
-                    "경기도", "안양시", "만안구", "56789", "2층"
-            });
-        }
-
-        jdbcTemplate.batchUpdate(memberSql, members, BATCH_SIZE,
-                (ps, param) -> {
-                    ps.setString(1, (String) param[0]);
-                    ps.setString(2, (String) param[1]);
-                    ps.setString(3, (String) param[2]);
-                    ps.setString(4, (String) param[3]);
-                    ps.setString(5, (String) param[4]);
-                    ps.setString(6, (String) param[5]);
-                    ps.setString(7, (String) param[6]);
-                    ps.setString(8, (String) param[7]);
-                    ps.setString(9, (String) param[8]);
-                });
-        Long myGreatBidding = 0L;
-
-        Bidding bidding = Bidding.builder()
-                .auction(auction)
-                .member(member)
-                .price(price)
-                .build();
-        ReflectionTestUtils.setField(bidding, "id", 2L);
-        when(biddingRepository.save(any(Bidding.class))).thenReturn(bidding);
-
-        List<Member> customers = memberRepository.findAll().stream()
-                .filter(m -> m.getMemberRole() == MemberRole.CUSTOMER)
-                .toList();
-        doNothing().when(paymentDepositService).payBiddingPrice(anyLong(), anyLong(), anyLong());
-
-        String depositSql = "INSERT INTO payment_deposit (member_id, deposit_amount) VALUES (?, ?)";
-        List<Object[]> paymentDeposits = new ArrayList<>();
-        // when
-        biddingService.createBidding(auctionId, price, authMember);
-
-        for (Member customer : customers) {
-            paymentDeposits.add(new Object[]{customer.getId(), 1_000_000L});
-        }
-        // then
-        assertEquals(price, auction.getCurrentPrice()); //  현재 가격이 업데이트되었는지 검증
-        assertEquals(price, bidding.getPrice());
-        assertNotNull(bidding);
-
-        jdbcTemplate.batchUpdate(depositSql, paymentDeposits, BATCH_SIZE,
-                (ps, param) -> {
-                    ps.setLong(1, (Long) param[0]);
-                    ps.setLong(2, (Long) param[1]);
-                });
-        verify(paymentDepositService, times(1)).payBiddingPrice(memberId, price, myGreatBidding);
-        verify(biddingRepository, times(1)).save(any(Bidding.class));
-    }
+    @Mock
+    private BiddingRepository biddingRepository;
+    @Mock
+    private AuctionRepository auctionRepository;
+    @Mock
+    private MemberRepository memberRepository;
+    @Mock
+    private KeyboardRepository keyboardRepository;
+    @Mock
+    private PaymentDepositService paymentDepositService;
+    @Mock
+    private PaymentDepositRepository paymentDepositRepository;
 
     @BeforeEach
-    public void setUp() {
-        setUpBidding();
+    public void setup() {
+        setUpTestData();
     }
 
+    private void setUpTestData() {
+        memberRepository.save(TestData.TEST_MEMBER1);
+        memberRepository.save(TestData.TEST_MEMBER2);
+        memberRepository.save(TestData.TEST_MEMBER3);
+        memberRepository.save(TestData.TEST_MEMBER4);
+        keyboardRepository.save(TestData.TEST_KEYBOARD1);
+        keyboardRepository.save(TestData.TEST_KEYBOARD2);
+        keyboardRepository.save(TestData.TEST_KEYBOARD3);
+        keyboardRepository.save(TestData.TEST_KEYBOARD4);
+        keyboardRepository.save(TestData.TEST_KEYBOARD5);
+        keyboardRepository.save(TestData.TEST_KEYBOARD6);
+        keyboardRepository.save(TestData.TEST_KEYBOARD7);
+        auctionRepository.save(TestData.TEST_AUCTION1);
+        auctionRepository.save(TestData.TEST_AUCTION2);
+        auctionRepository.save(TestData.TEST_AUCTION3);
+        auctionRepository.save(TestData.TEST_AUCTION4);
+        auctionRepository.save(TestData.TEST_AUCTION5);
+        auctionRepository.save(TestData.TEST_AUCTION6);
+        auctionRepository.save(TestData.TEST_AUCTION7);
+        biddingRepository.save(TestData.TEST_BIDDING1);
+        biddingRepository.save(TestData.TEST_BIDDING2);
+        biddingRepository.save(TestData.TEST_BIDDING3);
+        biddingRepository.save(TestData.TEST_BIDDING4);
+        paymentDepositRepository.save(TestData.TEST_PAYMENT_DEPOSIT1);
+        paymentDepositRepository.save(TestData.TEST_PAYMENT_DEPOSIT2);
+        paymentDepositRepository.save(TestData.TEST_PAYMENT_DEPOSIT3);
+        paymentDepositRepository.save(TestData.TEST_PAYMENT_DEPOSIT4);
+    }
+
+    //todo 성공케이스만들기
 
     @Test
-    @DisplayName("비관적 락 이용한 입찰 생성")
-    public void createBiddingWithPessimisticLock() throws InterruptedException {
-        List<Member> members = memberRepository.findAll(); // 전체 유저를 조회하고
-        // customer만 담기
-        List<Member> customers = new ArrayList<>();
-    void createBidding_success_두번째이상() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID1;
-        Long memberId = TestData.TEST_ID2;
-        Long price = 25000L; //  새로운 입찰가
-        Long myGreatBidding = TestData.TEST_AUCTION1.getCurrentPrice(); //  기존 입찰가
-
+    @DisplayName("진행중인 경매가 아닐때")
+    public void createBidding_fail_Auction_Not_Inprogress() {
+        Long auctionId = TestData.TEST_AUCTION2.getId();
+        Long memberId = TestData.TEST_MEMBER4.getId();
+        Long price = 25000L;
         AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
 
-        for (Member member : members) {
-            if (member.getMemberRole() == MemberRole.CUSTOMER) {
-                customers.add(member);
-            }
-        }
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-        ExecutorService executorService = Executors.newFixedThreadPool(10);
-        CountDownLatch latch = new CountDownLatch(100);
-        Auction auction = TestData.TEST_AUCTION1; //  실제 객체 사용
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        for (int i = 0; i < 100; i++) {
-            final long biddingPrice = auction.getCurrentPrice() + (i + 1) * 100L;
-            final Member customer = customers.get(i);
-        when(biddingRepository.findByMember_IdAndAuction_Id(memberId, auctionId))
-                .thenReturn(Optional.of(myGreatBidding)); //  기존 입찰 내역 존재
-
-            executorService.execute(() -> {
-                try {
-                    biddingService.createBidding(
-                            auction.getId(),
-                            biddingPrice,
-                            new AuthMember(customer.getId(), customer.getMemberRole())
-                    );
-                } catch (DataInvalidException e) {
-                    log.info("입찰 실패 - 유저 ID: {}, 현재가: {}, 입찰 금액: {}", customer.getId(),
-                            auction.getCurrentPrice(), biddingPrice);
-                    log.info("예외 메시지: {}", e.getMessage());
-                } finally {
-                    latch.countDown();
-                }
-            });
-        }
-        Bidding bidding = Bidding.builder()
-                .auction(auction)
-                .member(member)
-                .price(price)
-                .build();
-        ReflectionTestUtils.setField(bidding, "id", 2L);
-        when(biddingRepository.save(any(Bidding.class))).thenReturn(bidding);
-
-        latch.await();
-        executorService.shutdown();
-        doNothing().when(paymentDepositService).payBiddingPrice(anyLong(), anyLong(), anyLong());
-
-        // when
-        biddingService.createBidding(auctionId, price, authMember);
-
-        // then
-        // 최신 현재가
-        Auction updatedAuction = auctionRepository.findById(auction.getId()).get();
-        assertEquals(price, auction.getCurrentPrice()); //  현재 가격이 업데이트되었는지 검증
-        assertEquals(price, bidding.getPrice());
-        assertNotNull(bidding);
-
-        //  기존 입찰가가 `payBiddingPrice()`에서 정상적으로 반영되었는지 검증
-        verify(paymentDepositService, times(1)).payBiddingPrice(memberId, price, myGreatBidding);
-
-        //  새로운 입찰가로 저장되었는지 검증
-        verify(biddingRepository, times(1)).save(any(Bidding.class));
-    }
-
-
-    @Test
-    void createBidding_실패1_경매가_진행중이_아님() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID2; //  진행 중이 아닌 경매 사용
-        Long memberId = TestData.TEST_ID2;
-        Long price = 40000L;
-
-        AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
-
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-        Auction auction = TestData.TEST_AUCTION2; //  실제 객체 사용
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        // when, then
-        Exception exception = assertThrows(
-                OperationNotAllowedException.class, //  예외 타입 검증
-                () -> biddingService.createBidding(auctionId, price, authMember) //  실행
-        );
-
-        //  예외 메시지 검증
-        assertEquals(ErrorCode.AUCTION_NOT_IN_PROGRESS,
-                ((OperationNotAllowedException) exception).getErrorCode());
-    }
-
-        // 최종 입찰가
-        Bidding lastBidding = biddingRepository.findByAuctionIdOrderByPriceDesc(auction.getId())
-                .get(0);
-        assertEquals(lastBidding.getPrice(), updatedAuction.getCurrentPrice());
-    @Test
-    void createBidding_실패2_비딩_횟수가_10번이상() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID1;
-        Long price = 30000L;
-        Long memberId = TestData.TEST_ID2;
-
-        AuthMember authMember = new AuthMember(memberId, MemberRole.CUSTOMER);
-
-        //mock
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-        Auction auction = TestData.TEST_AUCTION1;
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        // 비딩 횟수 초과
-        when(biddingRepository.countByMember_IdAndAuction_Id(anyLong(), anyLong())).thenReturn(10L);
-
-        //when/then
-        Exception exception = assertThrows(
+        OperationNotAllowedException exception1 = assertThrows(
                 OperationNotAllowedException.class,
                 () -> biddingService.createBidding(auctionId, price, authMember)
         );
-
-        //메세지 검증
-        assertEquals(ErrorCode.MAX_BIDDING_COUNT_EXCEEDED,
-                ((OperationNotAllowedException) exception).getErrorCode());
+        assertEquals("진행 중인 경매가 아닙니다.", exception1.getErrorCode().getMessage());
     }
 
     @Test
-    void createBidding_실패3_비딩금액의_단위가_설정된_단위보다_작음() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID3;
-        ;
-        Long price = 45000L;
-        Long memberId = TestData.TEST_ID2;
+    @DisplayName("비딩횟수 초과")
+    public void createBidding_fail_bids_exceeded() {
+        Long auctionId = TestData.TEST_AUCTION3.getId();
+        Long memberId = TestData.TEST_MEMBER4.getId();
+        Long price = 25000L;
+        AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
 
-        AuthMember authMember = new AuthMember(memberId, MemberRole.CUSTOMER);
-        //mock
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-        Auction auction = TestData.TEST_AUCTION3;
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        //when, then
-        Exception exception = assertThrows(
-                DataInvalidException.class,
+        // 실제 DB에 10개의 비딩 추가
+        for (int i = 0; i < 10; i++) {
+            biddingRepository.save(Bidding.builder()
+                    .auction(TestData.TEST_AUCTION3)
+                    .member(TestData.TEST_MEMBER4)
+                    .price(price + (i * 100L))
+                    .build());
+        }
+        OperationNotAllowedException exception2 = assertThrows(
+                OperationNotAllowedException.class,
                 () -> biddingService.createBidding(auctionId, price, authMember)
         );
-
-        //메세지 검증
-        assertEquals(ErrorCode.INVALID_BIDDING_PRICE_UNIT,
-                ((DataInvalidException) exception).getErrorCode());
+        assertEquals("입찰은 10번까지만 가능합니다.", exception2.getErrorCode().getMessage());
     }
 
     @Test
-    void createBidding_실패4_현재가보다_낮은_입찰금액() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID3;
-        Long price = 30000L;
-        Long memberId = TestData.TEST_ID2;
+    @DisplayName("입찰단위에 맞지 않는 입찰")
+    public void createBidding_fail_bid_not_fit_unit() {
+        Long auctionId = TestData.TEST_AUCTION1.getId();
+        Long memberId = TestData.TEST_MEMBER4.getId();
+        Long price = 25100L;
+        AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
 
-        AuthMember authMember = new AuthMember(memberId, MemberRole.CUSTOMER);
-        //mock
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-        Auction auction = TestData.TEST_AUCTION3;
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        //when, then
-        Exception exception = assertThrows(
-                DataInvalidException.class,
+        DataInvalidException exception3 = assertThrows(DataInvalidException.class,
                 () -> biddingService.createBidding(auctionId, price, authMember)
         );
-
-        //메세지 검증
-        assertEquals(ErrorCode.BIDDING_PRICE_BELOW_CURRENT_PRICE,
-                ((DataInvalidException) exception).getErrorCode());
+        assertEquals("최소 입찰 금액 단위의 배수가 아닙니다.", exception3.getErrorCode().getMessage());
     }
 
     @Test
-    void createBidding_실패5_최대호가보다_높은_입찰금액() {
-        // given
-        Long auctionId = TestData.TEST_AUCTION_ID1;
-        Long price = 100000000L;
-        Long memberId = TestData.TEST_ID2;
+    @DisplayName("현재가보다 낮은 입찰금액")
+    public void createBidding_fail_lower_than_current_price() {
+        Long auctionId = TestData.TEST_AUCTION1.getId();
+        Long memberId = TestData.TEST_MEMBER4.getId();
+        Long price = 20000L;
+        AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
 
-        AuthMember authMember = new AuthMember(memberId, MemberRole.CUSTOMER);
-        //mock
-        Member member = TestData.TEST_MEMBER2;
-        ReflectionTestUtils.setField(member, "id", memberId);
-        when(memberRepository.findById(memberId)).thenReturn(Optional.of(member));
-
-        Auction auction = TestData.TEST_AUCTION1;
-        ReflectionTestUtils.setField(auction, "id", auctionId);
-        when(auctionRepository.findById(auctionId)).thenReturn(Optional.of(auction));
-
-        //when, then
-        Exception exception = assertThrows(
-                DataInvalidException.class,
+        DataInvalidException exception4 = assertThrows(DataInvalidException.class,
                 () -> biddingService.createBidding(auctionId, price, authMember)
         );
-
-        //메세지 검증
-        assertEquals(ErrorCode.BIDDING_PRICE_EXCEEDS_MAX_LIMIT,
-                ((DataInvalidException) exception).getErrorCode());
+        assertEquals("입찰가가 현재가보다 작습니다.", exception4.getErrorCode().getMessage());
     }
 
+    @Test
+    @DisplayName("최대호가보다 높은 입찰금액")
+    public void createBidding_fail_higher_than_max_price() {
+        Long auctionId = TestData.TEST_AUCTION1.getId();
+        Long memberId = TestData.TEST_MEMBER4.getId();
+        Long price = 100000L;
+        AuthMember authMember = new AuthMember(memberId, TestData.TEST_MEMBER_ROLE2);
+
+        DataInvalidException exception5 = assertThrows(DataInvalidException.class,
+                () -> biddingService.createBidding(auctionId, price, authMember)
+        );
+        assertEquals("입찰가가 1회 입찰 시 가능한 최대 금액(최소 입찰 금액 단위의 10배)을 초과하였습니다.",
+                exception5.getErrorCode().getMessage());
+    }
+
+    @Test
+    @DisplayName("입찰 조회")
+    public void getbidding_by_auction() {
+        Long auctionId = TestData.TEST_AUCTION1.getId();
+
+        List<BiddingResponseDto> result = biddingService.getBiddingByAuction(auctionId);
+
+        assertEquals(3, result.size());
+        assertEquals(26000L, result.get(0).getBiddingPrice());
+        assertEquals(24000L, result.get(1).getBiddingPrice());
+        assertEquals(21000L, result.get(2).getBiddingPrice());
+    }
+
+    @Test
+    @DisplayName("낙찰 조회")
+    public void getbidding_success_bidding() {
+
+    }
 }
+
